850: delete sample module - Sample module deleted - Moved docker build to server module
Issue: https://github.com/secureapigateway/secureapigateway/issues/850
