--- secure-api-gateway-ob-uk-rs-sample/README.md
@@ -1,2 +1,0 @@
-## secure-api-gateway-ob-uk-rs-sample
-A sample Spring Boot application to start up the UK Open Banking RS Simulator for the Secure API Gateway.
