--- secure-api-gateway-ob-uk-rs-sample/src/main/java/com/forgerock/sapi/gateway/ob/uk/rs/sample/BankSimulatorApplication.java
@@ -1,36 +1,0 @@
-/*
- * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.sapi.gateway.ob.uk.rs.sample;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@ComponentScan(basePackages =
-        {
-                "com.forgerock.sapi.gateway.common.cors",
-                "com.forgerock.sapi.gateway.ob.uk.rs"
-        }
-)
-@SpringBootApplication
-@EnableMongoRepositories(basePackages = "com.forgerock.sapi.gateway.ob.uk.rs.server.persistence.repository")
-public class BankSimulatorApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BankSimulatorApplication.class, args);
-    }
-}
