--- secure-api-gateway-ob-uk-rs-sample/src/test/resources/application-test.yml
@@ -1,20 +1,0 @@
-spring:
-  cloud:
-    config:
-      enabled: false
-
-  data:
-    mongodb:
-      # force SpringBoot tests to use a random port for mongo
-      port: 0
-
-rs:
-  discovery:
-    financialId: 0015800001041REAAY
-
-  # Data creation limits (see com.forgerock.sapi.gateway.ob.uk.rs.admin.api.data.DataCreator)
-  data:
-    upload:
-      limit:
-        accounts: 100
-        documents: 1000
