+++ secure-api-gateway-ob-uk-rs-server/README.md
@@ -1,3 +1,3 @@
 ## secure-api-gateway-ob-uk-rs-server
 
-A UK Open Banking RS simulator for the Secure API Gateway
+A UK Open Banking RS simulator server Spring boot application for the Secure API Gateway
