+++ secure-api-gateway-ob-uk-rs-server/docker/scripts/wait-for.sh
