+++ secure-api-gateway-ob-uk-rs-server/src/test/resources/application-test.yml
@@ -1,3 +1,4 @@
+#Spring config
 spring:
   cloud:
     config:
